FEAT: Added Authentication and Authorization

+++ src/main/java/com/example/airbnb/airbnb_app/service/RoomServiceImpl.java
@@ -3,12 +3,15 @@ package com.example.airbnb.airbnb_app.service;
 import com.example.airbnb.airbnb_app.dto.RoomDto;
 import com.example.airbnb.airbnb_app.entity.Hotel;
 import com.example.airbnb.airbnb_app.entity.Room;
+import com.example.airbnb.airbnb_app.entity.User;
 import com.example.airbnb.airbnb_app.exception.ResourceNotFoundException;
+import com.example.airbnb.airbnb_app.exception.UnAuthorisedException;
 import com.example.airbnb.airbnb_app.repository.HotelRepository;
 import com.example.airbnb.airbnb_app.repository.RoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -30,6 +33,12 @@ public class RoomServiceImpl implements RoomService{
         Hotel hotel = hotelRepository
                 .findById(hotelId)
                 .orElseThrow(() -> new ResourceNotFoundException("Hotel not found with Id: " + hotelId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(hotel.getOwner())) {
+            throw new UnAuthorisedException("This user does not own this hotel with id: "+hotelId);
+        }
+
         Room room = modelMapper.map(roomDto, Room.class);
         room.setHotel(hotel);
         room = roomRepository.save(room);
@@ -70,6 +79,11 @@ public class RoomServiceImpl implements RoomService{
         Room room = roomRepository
                 .findById(roomId)
                 .orElseThrow(() -> new ResourceNotFoundException("Room not found with ID: "+roomId));
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        if(!user.equals(room.getHotel().getOwner())) {
+            throw new UnAuthorisedException("This user does not own this room with id: "+roomId);
+        }
 
         inventoryService.deleteAllInventories(room);
         roomRepository.deleteById(roomId);
